tested GameService (#60)
* tested GameService

* refactored the repository layer and written integration test for the repo layer

+++ SteamStore/Repositories/CartRepository.cs
@@ -38,7 +38,7 @@ public class CartRepository: ICartRepository
                     Name = (string)row[SqlConstants.NAME_COLUMN ],
                     Description = (string)row[SqlConstants.DESCRIPTION_COLUMN],
                     ImagePath = (string)row[SqlConstants.IMAGE_URL_COLUMN],
-                    Price = Convert.ToDouble(row[SqlConstants.PRICE_COLUMN]),
+                    Price = Convert.ToDecimal(row[SqlConstants.PRICE_COLUMN]),
                     Status = "Approved"
                 };
                 games.Add(game);
